refactor: use kebab-case instead of camelCase
