+++ src/components/forms/font-select/font-select.tsx
@@ -1,5 +1,5 @@
 import { SingleValue } from 'react-select';
-import { GoogleFontsFamily, useGoogleFont } from '../../../hooks/fonts/useGoogleFont';
+import { GoogleFontsFamily, useGoogleFont } from '../../../hooks/fonts/use-google-font';
 import { convertFontWeight } from '../../../utils/fonts/convert-weight';
 import { Select } from '../select/select';
